Added concurrenci managment (Simulator.h and .c) edited stadistics recopilated on frame_allocator, paginacion and segmentacion
TODO:
Report.h and .c (for avoid overcharge the main)
workloads
stadistics for tlb

+++ include/tlb.h
@@ -8,7 +8,7 @@
 */
 
 typedef struct{
-    uint64_t frame_number; 
+    int frame_number; 
     int valid;
     int vpn; // page_number
 } tlb_entry;
